pyproject.toml
[tool.poetry]
name = "itch-dl"
packages = [{ include = "itch_dl" }]
version = "0.1.0"
description = "itch.io bulk game downloader"
homepage = "https://github.com/DragoonAethis/itch-dl"
repository = "https://github.com/DragoonAethis/itch-dl"
authors = ["Dragoon Aethis <dragoon@dragonic.eu>"]
readme = "README.md"
license = "MIT"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Games/Entertainment",
    "Topic :: Utilities",
    "Typing :: Typed"
]

[tool.poetry.urls]
"Wiki" = "https://github.com/DragoonAethis/itch-dl/wiki"
"Bug Tracker" = "https://github.com/DragoonAethis/itch-dl/issues"

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.64.0"
urllib3 = "^1.26.9"
requests = "^2.27.1"
python-slugify = "^6.1.2"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"

[tool.poetry.scripts]
itch-dl = "itch_dl.cli:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
